requirements.txt
asyncpg==0.31.0
aiofiles==25.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pytest==8.1.1
jupyter==1.0.0
pandas==2.2.3
matplotlib==3.8.4
seaborn==0.13.2
